refactor: change folder of shared
