+++ apps/web/src/services/dashboards/constants/widget-container-config.ts
