+++ apps/web/src/services/dashboards/helpers/__tests__/widget-width-helper.test.ts
@@ -2,11 +2,11 @@ import { describe, expect } from 'vitest';
 
 import type { WidgetSize } from '@/schema/dashboard/_types/widget-type';
 
-import { widgetWidthAssigner } from '@/services/dashboards/helpers/widget-width-helper';
 import {
     WIDGET_CONTAINER_MAX_WIDTH,
     WIDGET_CONTAINER_MIN_WIDTH, WIDGET_WIDTH_CRITERIA,
-} from '@/services/dashboards/shared/dashboard-widget-container/widget-container-config';
+} from '@/services/dashboards/constants/widget-container-config';
+import { widgetWidthAssigner } from '@/services/dashboards/helpers/widget-width-helper';
 
 
 /**
