+++ apps/web/src/services/dashboards/helpers/widget-theme-helper.ts
@@ -1,4 +1,4 @@
-import type { WidgetThemeOption, WidgetThemeAssignedList } from '@/services/dashboards/shared/dashboard-widget-container/widget-container-config';
+import type { WidgetThemeOption, WidgetThemeAssignedList } from '@/services/dashboards/constants/widget-container-config';
 import { WIDGET_THEMES } from '@/services/dashboards/widgets/_types/widget-type';
 
 const WIDGET_THEME_LEN = WIDGET_THEMES.length;
